feat: Add ChatInterface to ExercisesPager with fullscreen workspace
- Extract SplitPaneLayout from ExerciseWorkspace into reusable component
- Refactor ExerciseWorkspace to use SplitPaneLayout for split-pane behavior
- Rename pdfContent → primaryContent in ExerciseWorkspace (semantic cleanup)
- Add ChatInterface to ExercisesPager for exercise steps
- Use ExerciseWorkspace as host shell: intro/outro outside, exercises inside
- Make chat lesson-scoped (persists across exercise navigation)
- Fix SplitPaneLayout to not center content (exercise content fills pane)

Files:
- NEW: split-pane-layout.tsx (reusable split-pane component)
- MOD: ExerciseWorkspace/index.tsx (use SplitPaneLayout)
- MOD: ExercisesPager/index.tsx (use ExerciseWorkspace for exercises)
- MOD: lesson/[lessonSlug]/page.tsx (use primaryContent prop)
