+++ src/app/(frontend)/courses/[courseSlug]/chapters/[chapterSlug]/lessons/[lessonSlug]/_components/ExercisesPager/index.tsx
@@ -9,6 +9,8 @@ import { useTranslations } from '@/ui/web/providers/I18n'
 import type { ExerciseContentData } from '@/ui/web/exerciserenderer/types'
 import { Progress } from '@/ui/web/components/progress'
 import { useExercisesPager } from './useExercisesPager'
+import { ExerciseWorkspace } from '@/app/(frontend)/courses/[courseSlug]/chapters/[chapterSlug]/lessons/[lessonSlug]/exercises/[exerciseSlug]/_components/ExerciseWorkspace'
+import { ChatInterface } from '@/ui/web/chat'
 
 interface ExercisesPagerProps {
   exercises: Exercise[]
@@ -17,6 +19,7 @@ interface ExercisesPagerProps {
   courseSlug: string
   chapterSlug: string
   lessonSlug: string
+  lessonId: string
 }
 
 export function ExercisesPager({
@@ -26,6 +29,7 @@ export function ExercisesPager({
   courseSlug,
   chapterSlug,
   lessonSlug,
+  lessonId,
 }: ExercisesPagerProps) {
   const t = useTranslations('courses')
   const {
@@ -41,6 +45,75 @@ export function ExercisesPager({
   } = useExercisesPager({ exercises, courseSlug, chapterSlug, lessonSlug })
 
   const exerciseOrdinal = getExerciseOrdinal()
+  const currentExercise =
+    typeof pageState.exerciseIndex === 'number' ? exercises[pageState.exerciseIndex] : null
+
+  if (pageState.type === 'exercise' && currentExercise) {
+    return (
+      <ExerciseWorkspace
+        exerciseTitle={currentExercise.title}
+        backUrl={backUrl}
+        primaryContent={
+          <div className="h-full overflow-y-auto">
+            <div className="w-full p-4 md:p-6 space-y-4">
+              <Progress value={progressPercent} className="h-1 rounded-full" />
+
+              <div className="bg-card rounded-2xl p-5 md:p-6 border border-border/60 shadow-sm">
+                <div className="flex items-center gap-3 mb-2">
+                  <div className="w-8 h-8 bg-primary/10 rounded-lg flex items-center justify-center">
+                    <Layers className="w-4 h-4 text-primary" />
+                  </div>
+                  <div>
+                    <p className="text-[10px] text-muted-foreground uppercase tracking-[0.15em]">
+                      {exerciseOrdinal !== null
+                        ? `${t('exercise')} ${exerciseOrdinal} ${t('of')} ${totalExercises}`
+                        : ''}
+                    </p>
+                    <h2 className="text-lg font-medium text-foreground">{currentExercise.title}</h2>
+                  </div>
+                </div>
+              </div>
+
+              <div className="bg-card rounded-2xl p-5 md:p-6 border border-border/60 shadow-sm">
+                <ExerciseRenderer
+                  content={currentExercise.content as unknown as ExerciseContentData}
+                  mode="student"
+                  showCheckAnswer={true}
+                />
+              </div>
+
+              <div className="flex justify-between items-center pt-2">
+                <Button
+                  variant="ghost"
+                  onClick={handlePrev}
+                  disabled={!canGoPrev}
+                  className="text-muted-foreground text-sm hover:text-foreground gap-1.5"
+                >
+                  <ChevronRight className="w-4 h-4 rtl:rotate-0 ltr:rotate-180" />{' '}
+                  {t('exercisesPagerPrev')}
+                </Button>
+                <Button
+                  onClick={handleNext}
+                  disabled={!canGoNext}
+                  className="px-6 py-2 rounded-xl text-sm"
+                >
+                  {t('exercisesPagerNext')}
+                </Button>
+              </div>
+            </div>
+          </div>
+        }
+        chatContent={
+          <ChatInterface
+            lessonId={lessonId}
+            exerciseId={currentExercise.id}
+            translationNamespace="courses"
+            showMathTools={true}
+          />
+        }
+      />
+    )
+  }
 
   return (
     <div className="min-h-screen bg-background flex flex-col">
@@ -88,63 +161,6 @@ export function ExercisesPager({
                 >
                   {t('exercisesPagerStart')}{' '}
                   <ChevronLeft className="w-5 h-5 ms-2 rtl:rotate-0 ltr:rotate-180" />
-                </Button>
-              </div>
-            </div>
-          )}
-
-          {pageState.type === 'exercise' && typeof pageState.exerciseIndex === 'number' && (
-            <div className="space-y-8">
-              <div className="bg-card rounded-3xl p-6 md:p-8 border border-border/60 shadow-lg shadow-muted/40 relative overflow-hidden">
-                <div className="absolute top-0 end-0 w-1.5 h-full bg-primary rounded-s-full" />
-                <div className="flex items-center gap-3 mb-3">
-                  <div className="w-10 h-10 bg-primary/10 rounded-xl flex items-center justify-center">
-                    <Layers className="w-5 h-5 text-primary" />
-                  </div>
-                  <div>
-                    <p className="text-[10px] text-muted-foreground uppercase tracking-[0.15em]">
-                      {exerciseOrdinal !== null
-                        ? `${t('exercise')} ${exerciseOrdinal} ${t('of')} ${totalExercises}`
-                        : ''}
-                    </p>
-                    <h2 className="text-xl font-medium text-foreground">
-                      {exercises[pageState.exerciseIndex]?.title ?? ''}
-                    </h2>
-                  </div>
-                </div>
-              </div>
-
-              <div className="bg-card rounded-3xl p-6 md:p-8 border border-border/60 shadow-lg shadow-muted/40">
-                {typeof pageState.exerciseIndex === 'number' &&
-                  exercises[pageState.exerciseIndex] && (
-                    <ExerciseRenderer
-                      content={
-                        exercises[pageState.exerciseIndex]!
-                          .content as unknown as ExerciseContentData
-                      }
-                      mode="student"
-                      showCheckAnswer={true}
-                    />
-                  )}
-              </div>
-
-              <div className="flex justify-between items-center pt-4">
-                <Button
-                  variant="ghost"
-                  onClick={handlePrev}
-                  disabled={!canGoPrev}
-                  className="text-muted-foreground text-sm hover:text-foreground transition-colors duration-300 gap-1.5"
-                >
-                  <ChevronRight className="w-4 h-4 rtl:rotate-0 ltr:rotate-180" />{' '}
-                  {t('exercisesPagerPrev')}
-                </Button>
-                <Button
-                  onClick={handleNext}
-                  disabled={!canGoNext}
-                  size="lg"
-                  className="px-10 py-4 rounded-2xl text-base shadow-lg shadow-primary/20 hover:shadow-xl transition-all duration-300"
-                >
-                  {t('exercisesPagerNext')}
                 </Button>
               </div>
             </div>
