+++ src/app/(frontend)/courses/[courseSlug]/chapters/[chapterSlug]/lessons/[lessonSlug]/exercises/[exerciseSlug]/_components/ExerciseWorkspace/exercise-workspace-utils.ts
@@ -1,9 +1,7 @@
-import type { ViewMode } from './exercise-workspace-types'
-
 /**
  * Get initial view mode on page load
  * Always returns 'PDF' per HLS/GAP requirement
  */
-export function getInitialViewMode(): ViewMode {
+export function getInitialViewMode(): 'PDF' | 'CHAT' {
   return 'PDF'
 }
